requirements.txt
boto3==1.12.41
botocore==1.15.41
certifi==2019.6.16
chardet==3.0.4
decorator==4.4.0
Django==2.1.5
django-cors-headers==2.4.0
django-enumchoicefield==1.0.2
django-filter==2.1.0
djangorestframework==3.9.1
docutils==0.15.2
idna==2.8
imageio==2.5.0
imageio-ffmpeg==0.3.0
jmespath==0.9.5
Markdown==3.0.1
moviepy==1.0.0
numpy==1.16.4
Pillow==6.1.0
proglog==0.1.9
python-dateutil==2.8.1
pytz==2018.9
requests==2.22.0
s3transfer==0.3.3
six==1.14.0
tqdm==4.32.2
urllib3==1.25.3
